Simplified collide and stream
Updated the rest of the ("non-basic") functions to reflect this code simplification (there were some semi-significant changes to the initial setup). The boundary reflection works with the new implementation. The "basic" functions throw an indexing-related error I was not able to figure out.

Also updated the vector-reversing method.

Also cleaned the code up a little bit (names, syntax, simplifications).

The simulation is stable for the first few seconds after which it breaks down due to significant jumps in velocity which prompt priority investigation.

Performance is also down 20% on my end, but I was unable to find the reason for this.